refactor: Categorize components
