+++ src/components/blocks/EmbedBlock.vue
@@ -5,7 +5,7 @@
 <script lang="ts">
 import Vue from 'vue'
 import { EmbedBlock } from '../../types/Blocks'
-import ShadowInput from '../utils/ShadowInput.vue'
+import ShadowInput from '../blocks/ShadowInput.vue'
 import {
   isYouTubeVideoURL,
   isFacebookPostURL,
