+++ src/components/blocks/ImageBlock.vue
@@ -79,7 +79,7 @@ import Vue from 'vue'
 import axios from 'axios'
 import { ImageBlock } from '../../types/Blocks'
 import { createBlogImageFromDataURI } from '../../utils/createImage'
-import ShadowInput from '../utils/ShadowInput.vue'
+import ShadowInput from '../blocks/ShadowInput.vue'
 import AlignLeft from '../vector/AlignLeft.vue'
 import AlignCenter from '../vector/AlignCenter.vue'
 import AlignRight from '../vector/AlignRight.vue'
