+++ src/components/menu/InsertButtonIcon.vue
