+++ tests/unit/components/menu/InsertButton.spec.ts
@@ -1,5 +1,5 @@
 import { mount } from '@vue/test-utils'
-import InsertButton from '@/components/utils/InsertButton.vue'
+import InsertButton from '@/components/menu/InsertButton.vue'
 
 describe('InsertButton.vue', () => {
   test('renders correctly', () => {
