+++ tests/unit/components/menu/InsertPopup.spec.ts
@@ -1,5 +1,5 @@
 import { mount } from '@vue/test-utils'
-import InsertPopup from '@/components/utils/InsertPopup.vue'
+import InsertPopup from '@/components/menu/InsertPopup.vue'
 
 describe('InsertPopup.vue', () => {
   test('renders correctly', () => {
