+++ tests/unit/components/menu/InsertPopupIcon.spec.ts
@@ -1,5 +1,5 @@
 import { mount } from '@vue/test-utils'
-import InsertPopupIcon from '@/components/utils/InsertPopupIcon.vue'
+import InsertPopupIcon from '@/components/menu/InsertPopupIcon.vue'
 
 describe('InsertPopupIcon.vue', () => {
   test('renders correctly', () => {
